🚚 refactor: 将原本的 settings 更名为 global
