+++ src/pages/settings/ThemeSwatches/ThemeSwatchesNeutral.tsx
@@ -6,8 +6,7 @@ import {
   neutralColorsSwatches,
 } from '@lobehub/ui';
 import { memo } from 'react';
-
-import { useSettings } from '@/store/settings';
+import { useSettings } from 'src/store/global';
 
 const ThemeSwatchesNeutral = memo(() => {
   const [neutralColor, setSettings] = useSettings((s) => [s.settings.neutralColor, s.setSettings]);
