+++ src/pages/settings/layout.tsx
@@ -2,10 +2,10 @@ import Head from 'next/head';
 import { ReactNode, memo, useEffect } from 'react';
 import { useTranslation } from 'react-i18next';
 import { Flexbox } from 'react-layout-kit';
+import { useSettings } from 'src/store/global';
 
 import SideBar from '@/features/SideBar';
 import { createI18nNext } from '@/locales/create';
-import { useSettings } from '@/store/settings';
 import { genSiteHeadTitle } from '@/utils/genSiteHeadTitle';
 
 const initI18n = createI18nNext('setting');
