+++ src/store/global/index.ts
