+++ src/store/global/selectors.ts
